requirements.txt
wheel
getmac
requests
pycrypto
pypiwin32
setuptools
pycryptodome
python-dotenv

#@guthmaer